Encode/decode INI properties in IniFileChain since the encoding/decoding is used to avoid issues w/ parse_ini_file() use.
