+++ tests/PHPUnit/Unit/Config/IniFileChainTest.php
@@ -141,7 +141,7 @@ class IniFileChainTest extends PHPUnit_Framework_TestCase
                         )
                     ),
                     'Section2' => array(
-                        'var4' => 'value5'
+                        'var4' => 'val$ue5'
                     )
                 )
             ),
@@ -160,7 +160,7 @@ class IniFileChainTest extends PHPUnit_Framework_TestCase
                         )
                     ),
                     'Section2' => array(
-                        'var4' => 'value5'
+                        'var4' => 'val$ue5'
                     )
                 )
             )
@@ -194,7 +194,7 @@ class IniFileChainTest extends PHPUnit_Framework_TestCase
 
         $data =& $fileChain->get('Section1');
 
-        $this->assertEquals(array('var1' => 'value2', 'var3' => array('value3', 'value4')), $data);
+        $this->assertEquals(array('var1' => 'val"ue2', 'var3' => array('value3', 'value4')), $data);
 
         $data['var1'] = 'changed';
         $data['var3'][] = 'newValue';
@@ -241,7 +241,17 @@ class IniFileChainTest extends PHPUnit_Framework_TestCase
             __DIR__ . '/test_files/default_settings_2.ini.php'
         );
 
-        $this->assertEquals(array('var1' => 'value2', 'var3' => array('value3', 'value4')), $fileChain->getFrom($defaultSettingsPath, 'Section1'));
+        $this->assertEquals(array('var1' => 'val"ue2', 'var3' => array('value3', 'value4')), $fileChain->getFrom($defaultSettingsPath, 'Section1'));
+    }
+
+    public function test_getFrom_CorrectlyReturnsUnencodedValue()
+    {
+        $userSettingsPath = __DIR__ . '/test_files/special_values.ini.php';
+        $fileChain = new IniFileChain(array(), $userSettingsPath);
+
+        $this->assertEquals(array(
+            'value1' => 'a"bc', 'value2' => array('<script>', '${@piwik(crash))}'
+        )), $fileChain->getFrom($userSettingsPath, 'Section'));
     }
 
     public function getTestDataForDumpTest()
@@ -253,7 +263,7 @@ class IniFileChainTest extends PHPUnit_Framework_TestCase
                 ),
                 __DIR__ . '/test_files/default_settings_2.ini.php', // user settings
                 "; some header\n",
-                "; some header\n[Section1]\nvar1 = \"overriddenValue1\"\nvar3[] = \"overriddenValue2\"\nvar3[] = \"overriddenValue3\"\n\n[Section2]\nvar4 = \"value5\"\n\n",
+                "; some header\n[Section1]\nvar1 = \"overriddenValue1\"\nvar3[] = \"overriddenValue2\"\nvar3[] = \"overriddenValue3\"\n\n[Section2]\nvar4 = \"val&#36;ue5\"\n\n",
                 "; some header\n[Section1]\nvar1 = \"overriddenValue1\"\nvar3[] = \"overriddenValue2\"\nvar3[] = \"overriddenValue3\"\n\n"
             )
         );
