+++ tests/PHPUnit/Unit/Config/test_files/default_settings_1.ini.php
@@ -1,9 +1,9 @@
 [Section1]
 
-var1 = "value2"
+var1 = "val&quot;ue2"
 var3[] = "value3"
 var3[] = "value4"
 
 [Section2]
 
-var4 = "value5"
+var4 = "val&#36;ue5"
